requirements.txt
flask
python-dotenv
langchain
openai
groq
langchain_community
langchain_groq
